Provide downloadurl and web again in the reply
For some reason this got lost in the past, thus no automatic download
would occur anymore for the updates. Instead a browser pointing to the
payload would be open incurring lots of manual steps for the user.

With downloadurl this becomes automatic again, but it requires the
client to be able to follow redirects which it currently can't (fix
pending in the client for that aiming at 3.0.3).

Signed-off-by: Kevin Ottens <kevin.ottens@nextcloud.com>