Remove the aggressive throttling
Since nowadays most clients in the wild are 3.0.3 or above we should
likely remove that random roll of wrongly claiming the client is up to
date and rely on the update segment only again.

I'd expect this would fix nextcloud/desktop#2809 we're getting too many
of those "why didn't I get the update yet" and that's the only
explanation I have so far.

Signed-off-by: Kevin Ottens <kevin.ottens@nextcloud.com>